Moved format json converter tests to data module
