+++ data/src/test/java/com/aconno/sensorics/data/mapper/AconnoAdvertisementFormats.kt
@@ -1,4 +1,4 @@
-package com.aconno.sensorics.device.format
+package com.aconno.sensorics.data.mapper
 
 object AconnoAdvertisementFormats {
