--- device/src/test/java/com/aconno/sensorics/device/format/FormatJsonConverterImplTest.kt
@@ -1,26 +1,0 @@
-package com.aconno.sensorics.device.format
-
-import org.junit.Test
-
-class FormatJsonConverterImplTest {
-
-//    private val formatJsonConverter =
-//        FormatJsonConverterImpl(
-//            GsonBuilder()
-//                .setFieldNamingPolicy(FieldNamingPolicy.LOWER_CASE_WITH_UNDERSCORES)
-//                .create()
-//        )
-
-    @Test
-    fun convert_acnSensaScalarFormat() {
-//        val advertisementFormat =
-//            formatJsonConverter.toAdvertisementFormat(AconnoAdvertisementFormats.ACN_SENSA_SCALAR)
-//
-//        assertEquals("CF01", advertisementFormat.id)
-//        assertEquals("AcnSensa", advertisementFormat.getName())
-//        assertEquals("ic_sensa", advertisementFormat.getIcon())
-//
-//        assertEquals(5, advertisementFormat.getFormat().size)
-//        assertEquals(5, advertisementFormat.getRequiredFormat().size)
-    }
-}
